Move model files to its own package
